Refactor DetectorDummyPlugin to implement PupilDetectorPlugin interface
